Ensure all plugins have top-level name
Each plugin part must check if any previous part
resulted in an entry for their parent plugin being
added to the master yaml and if not add it
themselves. We also now ensure that plugins only
produce an entry in the master yaml if they actually
have information to show.