Add log parser to ceph plugin
Adds the start of a new part for the storage plugin that
searches Ceph daemon logs for events that indicate
something has gone wrong.

Also adds support to FileSearcher for providing a search
hint that is executed as a fast pre-search to eliminate
excessive full searches and hopefully save time.