Add known_bugs utils
This allows each plugin to collect a set of known bugs in
a temporary yaml file that is then added to the master
yaml as the last part of a plugin to be executed. Adds a
99known_bugs part to each plugin.

The change also includes a refactor of how we manage plugin
yaml and the shared code to perform those actions. It also
enables some relevant pylint cases that were needed to
catch issues previously undetected.