Use json output for ceph commands
The `plain` output format is fragile when it comes to output changes
such as between Stein and Train: Train adds another column which
changes the position of the PGS column. The `json_outout` is
straightforward to interpret on the other hand. This change adds a
function to read the json formatted output.

This change also updates the pylint configuration file to allow for
longer strings to allow for the additional artifacts in the storage
tests.

Signed-off-by: Nicolas Bock <nicolas.bock@canonical.com>