Add support for minimal output
By default hotsos runs all plugins and shows the full output
of data collected. For some cases in might be more helpful
to only show bugs and issues discovered along the way,
allowing optional further analysis if required. To achieve
this we now have a --short option that, if provided, will
filter the output to only include "known-bugs" and
"potential-issues" collections for each plugin that
has populated them with a value.

This patch also contains the code to support
potential-issues-found and adds an initial instance of this
to the kernel plugin.