Make plugins python importable
Python cannot (natively) import files whose name starts with
a number. This resulted in a lot of ugly hackery to allow plugin
parts to be imported into unit tests. This patch renames plugin
parts to start with a "_" and end in .py if they are python
(which at present they all are) so that they can be directly
imported into unit tests.