Extend sequence searching support
Adds a new class SequenceSearchDef to common.searchtools that
can be used to identify and extract sequences from files that
match a start and end expression with an option to also
extract information between start and end. A good example
of such a file is a rabbitmqctl report so the rabbitmq
plugin now implements this class to extract basic information.