Remove dependency between plugin parts yaml
Code that relies on the output of other parts is bound to
break since we no longer guarantee the order in which
parts are executed (only the order their output is
included in the master yaml). Therefore we need a way for
parts to have that information first hand. To do this we
create a new base class OpenstackChecksBase and share it
between the vm_info and networks openstack parts.