Add filters support to searchtools
This allows a filter to be defined for a data source
as a means of pre-determining which lines we want to
include/exclude from full searches which can be very
costly to execute. Filters have been applied to the
openstack plugin to reduce the amount of time they
take to execute when run on large quantities of
logs.