Sort dir contents on filename
When providing a directory as path to searchtools
we now order files to be able to execute searches
and use filters to stop after a predefined point e.g.
date.

Also adds a --max-logrotate-depth option which
defaults to 7 and caps the depth of logrotated
copies of a logfile that we will search (when using
--all-logs mainly).