[openstack] extend router update checks
Router updates have a unique_id so we should use that as
the event id instead of the router uuid. We now also include
the router id in the results as well as reporting the number
of incomplete updates. Lastly, event duration is now purely
calculated from end - start rather than extracting from the
log.