[juju] fix pre/post juju 2.9.x support
As of Juju 2.9.x there are no longer per-unit systemd
services and therefore no unit processes visible in ps
which we were using to identify which charm units were
running or not. This patch fixes this by obtaining the
list of running units from the machine agent.conf while
maintaining legacy support for older versions of juju
that didnt have that information.