[storage] add ApplicationBase class
The plugin base class now inherits from this new class inside
which we define characteristics common to any application. The
first characteristic we define is bind_interfaces which an
application plugin should implement to reflect any interfaces
the application may be bound to.

The first implementation of this is provided for ceph.