[openstack] change filters for exception searches
Filtering ERROR, WARNING and Traceback for all
exception types can get very time consuming. Exceptions are
almost always ERRORs so we now only filter ERROR, WARNING.
We also move warning and error searches to their own
category and filter accordingly which seems to massively
reduce search times for logs with very large numbers of
warnings (that we are not interested in for exception
searches).