Ensure sos commands run with sos TZ
The CLIHelper will sometimes execute binaries locally when
producing output from a sosreport command e.g. journalctl
and in doing so it must run using the same timezone as the
sosreport host in order to produce the same output. This add
TZ to the environment of commands executed by the BinFileCmd
to ensure they produce a consistent result no matter what the
tz of the host that hotsos is being from.