Make plugin runtime conditional
All implementations of plugintools.PluginPartBase must
now implement the plugin_runnable property which must
return True if the part would like to be executed. This
is to allow plugins to decide at runtime whether to
execute based on e.g. whether the application is installed
and therefore avoid unnecessary plugin executions.