[storage] Detect if there are too many osdmaps
Mons store a number of osdmaps (default = 500). But this can
get arbitrarily large if the cluster wasn't in healthy state
for a while. These are trimmed down automatically *if* the
cluster comes back to healthy state. However, too many
maps in itself can be a problem and can also indicate that
trimming isn't happening due to bugs.

Closes: #131

Signed-off-by: Ponnuvel Palaniyappan <pponnuvel@gmail.com>