[core] fix cli_helpers and improve host_helpers
If a CLIHelper object is going to be re-used to run
the same command more than once it needs to reset
the state of the command so that it runs afresh each
time. This resolves a bug with host_helpers retreiving
ip addr from multiple namespaces and also allows us to
group all namespace searches into a single search.