Improve how generate short summary
Currently generating a --short summary means first
generating a full summary then filtering it. If we
wanted a full and short summary in one go we should
not need to run the analysis twice. This patch
introduces support for providing a full summary as
input with --user-summary which, when used in conjunction
with --short will provide a short summary of the original.

We also provide a new --full option that when specified
with --short will result in both a full and short summary
being generated. This only applies when --save is used.

Resolves: #163