Support different input types for yaml checks
Currently we only support using the yaml based definitions
for file-based searches. This patch introduces the ability
to define different data sources i.e. file or command. This
is now applied to all defs types i.e. events, bugs etc.

Also adds support for using both simple and seqeunce searches
in yaml defs. This means that searches can be specified as
a single search expression, a sequence with start, body and
optional end or a search with start and end that can then be
fed into core.analysis.LogEventStats.

Event searches that use commands are now ported to yaml.