[storage] Reduce osdmaps limit
The 10x limit seems pretty high. We have seen problems even
with less 5000 osdmaps as there could be lots of old maps
which may be permanently stuck without ever being pruned
by the bugs (in the message description). So it's better
if we simply warn as soon as it goes over the limit so that
it can be looked at to decide whether it's a serious issue.

Signed-off-by: Ponnuvel Palaniyappan <pponnuvel@gmail.com>