Support defining scenarios in yaml
This patch adds a new way to define checks in yaml called
"scenarios". Similar to existing defs it uses a yaml tree
to define a set of checks and conclusions based on the
outcome of those checks. Conclusions can be prioritised so
that the most appropriate one is chosen if multiple are
drawn.

In order to allow definitions to scale properly, a new way
to organise the files is implemented here that uses a
file/dir structure for defintions instead of having
everything in a single file.

The first scenario implemented is for flapping ceph-mon
elections based on a combination of log entries and network
interface errors.