Simplify yaml def requirements
This patch removes the "context" property and replaces the use
of the settings property in bugs defs with requires by extending
the requires property apt type to support a list of version
ranges. This therefore replaces the needs for "versions-affected"
settings and means that all handler can now make use of this
functionality.