Summary implementation improvements
Code within the plugins directory is now mainly used to
print summary information. This patch provides a cleaner
way to implement summary code by providing a way to
control the exact order in which entries are printed in
the output and using a standard naming for methods that
are used to produce output. Each plugin has a "summary.py"
that is the implementation of the summary output for that
plugin and a special method name of __summary_<key> is used
to identify entries for the summary along with the key under
which it will be printed.