Support machine readable output
Provides a new cli option --machine-readable which when
provided will ensure that output is not formatted for
humans (i.e. for readability). This option will mean
that e.g. potential-issues and known-bugs are output
as dictionaries (in yaml or json depending on format)
and this also means that extra fields can be exposed.
One new field 'context' is provided to allow extra context
to be associated with issues and bugs raised such as
information about which file and linenumber the issue was
found.

Partially-Resolves: #310