Organise existing scenario definitions
This patch tidies up the set of scnarios we have today so that
it is more obvious what they do and when/where they run. Also
adds directory global requirements for storage/ceph/ceph-{mon,osd}
so that the scenarios beneath only run if the corresponding apt
package is installed and updates the unit tests is_def filter
to support running directory globals as part of unit test runs.