Ensure scenarios only run if requirements met
The scenarios handler supports skipping scenarios
of they have a requirement defined and it is not
met but we were still executing some properties
unnecessarily. This patch ensures that all
scenarios bound by a requirement are fully
skipped.

Also updates the requirements for ceph-mon and
ceph-osd scenarios to only run if their eponymous
systemd service exists and is enabled.