Create generic implementation for logical collections
Some ydef properties (YPropertyRequires,
YPropertyConclusion) can define a set of other
properties on which their outcome if derived and these
can grouped in different ways e.g. by logical operator
that is applied to their results. This provides a common
way implement this logic for any property that wants to
use it.