Fix cryptography dependency for snap builds
Since the python cryptography module>3.4.8, it is obligatory to
depend on the Rust compiler.
This brings further dependencies on other libraries for the snap
and did not allowed for hotsos to run correctly.

This patch aims to pin cryptography to 3.4.8, and introduce build
options to exclude the rust compiler dependency.

Reverts: https://github.com/canonical/hotsos/pull/327
Closes: https://github.com/canonical/hotsos/issues/326

Signed-off-by: David Negreira <david.negreira@canonical.com>