Add http request info to summary
This adds a new event def that adds http GET|PUT|POST|DELETE
counts to the output summary. We only look at the most recent
logfile so as not to overwhelm the summary with informaton.
Currently implementation is watching for neutron http
requests but others can easily be added.

Also fixes a bug with event callbacks to ensure that when
methods are registered they can only be used by the event
they are linked to.