Fix Openstack router spawn event check
We were using an end expression that matches multiple times through
the life of a router but it is only actually the first occurence
that represents a spawn (the rest are just updates). So now we
match on the actual command used to spawn the keepalived process. We
also ensure the hotsos.core.analytics closes the loop.