Apply constraints to results by default
When search results are using in scenarios their constraints,
if provided, will be applied to the results by default so that
e.g. when they are referenced by PropertyCacheRefResolver they
will take the constraints into account. Therefore the cached
'results' will have constraints applied and a new cached
'results_raw' will be pre-constraints.