Break up scenarios/kernel/kernlog_checks.yaml
The call trace scenarios rely on priority to ensure
that the generic call trace conclusion is ignored if
other types are matches. This was breaking other
conclusions in the same file that did not specify a
priority so we now break the networking (non-calltrace)
scenarios into their own network.yaml file.