New format for saved output
If requested to save the output to disk, we now
save all supported formats (yaml, json) and styles
(short, full etc) beneath a directory. At the root
of the directory there is a symbolic link to the
hitherto common full yaml summary. This means that
to get different styles or formats won't require
re-running hotsos.