Avoid unbounded journalctl searches
This ensures that if journalctl is called without a
value for --since it will use a default one that
takes into account the ALL_LOGS and MAX_LOGROTATE_DEPTH
config options. This will enforce the same limits that
we use for logfile searches and therefore reduce load
incurred from unbound journal searches.

Resolves: #386