Introduce filesearcher constraints
Adds support for setting search constraints in the
filesearcher. When a search expression is registered
it can also have one or more constraints to define
the limitations or parameters of the search. Currently
we support a SearchSince constraint which supports
skipping log file lines that are not on or after a set
datetime (a delta from the current date). This can also
be applied globally to all searches. To achieve this
with minimal impact a binary pre-search is performed on
the file to seek to a valid start point. This will
avoid searches on data that is not within a required
time interval. If valid datetimes cannot be matched in
the file the constraint is not applied.