Fix config check issue message
The config check property will store the last key/ops/actual_value
to be checked and the bcache store checks where always reporting the
last config checked when raising an issue which was not necessarily
the one that failed its assertion. This fixes that problem by
introducing a new cache key 'assertion_results' that contains all
assertions so that can be included in a message.

Resolves: #393