Reduce ceph-mon tests mocking of CLIHelper
Mocking CLIHelper is sometimes necessary but it also
bloats the test code and has the risk of detaching
the code from the real CLIHelper implementation. The
preferred approach is to populate a custom DATA_ROOT
with whatever the test needs and have the real
CLIHelper consume that in the same way it would in
reality.