[plugin/ceph] Fine tune suboptimal PG count warning
The suboptimal PGs per PSD warning can at times be false.
A specific case is when a cluster is newly deployed. In
that case, there wouldn't much data, and so, either the
admin hasn't scaled the PGs yet or if the PG autoscaler
is enabled, then it doesn't think PGs need to be scaled
for the current utilisation.

So we now warn on suboptimal PGs if:
  - autoscaler isn't enabled, and
  - *every* pool in the cluster is utilised under 2%

Fixes #404.

Signed-off-by: Ponnuvel Palaniyappan <pponnuvel@gmail.com>