Introduce support for vars in ydefs.
Adds a new property "vars" that supports declaring a list
of variables that can be referenced in requirement properties
using a new "varops" type which takes a list of ops where
the first must be a variable to emulate the input. Variables
can also be used in a raises.format_dict. See unit test for
examples. Supports declaring variables with raw values or
imported python properties.

Adapts some existing scenario definitions to use variabled in
order to simplify their implementaton.