juju and ovs event/scenario tuning
Uses less permissive default input path for juju events. Since
juju log files tend to be large and don't logrotate they result
in lost of unnecessary searching. This sets unit log files as
the default and sets machine logs for checks that need them.

For openvswitch this fixes the unreasonably-long-poll-interval
search expressions which were erroneously including extra
information and therefore breaking the results processing
resulting in fields included in the output that were not
json-serialisable.