Remove shared-memory plug
The plug added in commit 047369c actually breaks uploads to
the snapstore since it wont allow classic snaps with
interfaces. The good news is that it seems with the latest
snapd we no longer need this or the preload hack for /dev/shm
to be accessible/usable so getting rid of both from now on.