Fix snap to support cryptography module
Because of `classic` confinement the dynamic linker will load the host's
`libc.so.6` instead of the library from the `core22` snap. This goes unnoticed
as long as the host running the snap is Focal. However, `glibc` versions more
recent than Focal do not include a private symbol which is needed by the
`cryptography` rust module and the following error will be thrown:

    ImportError: /snap/core22/current/lib/x86_64-linux-gnu/libpthread.so.0: undefined symbol: __libc_siglongjmp, version GLIBC_PRIVATE

In order to understand what is going on here, note that (1) `strict` snaps
create a filesystem namespace and mount the base snap so that a path such as
`/lib/x86_64-linux-gnu/libc.so.6` resolves to the library in the core snap and
not to the library on the host, and (2) `snapcraft` uses `patchelf` in `classic`
snaps to insert an `RPATH` into libraries so that the dynamic linker links to
the correct path at runtime since the core snap is not mounted into the snap
filesystem namespace.

In detail, when loading the cryptography rust module, the linker will load
`libpthread.so` from `core22` (which is correct since our snap base is `core22`)

    $ ldd /snap/hotsos/current/lib/python3.10/site-packages/cryptography/hazmat/bindings/_rust.abi3.so
            libpthread.so.0 => /snap/core22/current/lib/x86_64-linux-gnu/libpthread.so.0 (0x00007f5a7f5b9000)

This is then resolved to `libc.so.6` because:

    $ objdump -x /snap/core22/current/lib/x86_64-linux-gnu/libpthread.so.0 | grep libc
      NEEDED               libc.so.6
    0000000000000000       F *UND*  0000000000000000              __libc_siglongjmp@@GLIBC_PRIVATE

But

    $ ldd /snap/core22/current/lib/x86_64-linux-gnu/libpthread.so.0
            libc.so.6 => /lib/x86_64-linux-gnu/libc.so.6 (0x00007f5d67000000)

Which loads the host libc library.

In order to address this issue we need to include Python and libc6 in the hotsos
snap so they can be properly patched with `patchelf`. In addition, this changes

1. Cleans up the parts section
2. Switches to the `dump` plugin since the `python` plugin behaves incorrectly
   for our particular use case
3. Add explicit `patchelf` part
   (This part is not necessary for `core20` but will be required for
   `core22`.)
4. Generalize shebang regex
   On Launchpad things are buit in `/build` and locally in `/root`.
   Generalize the regular expression used by `sed` to account for this
   difference.
5. Make sure to also install `dist-info`
   The `install` command of `setuptools` does not install `dist-info`.
   By taking the scenic route and instead calling `bdist_wheel` and then
   `pip` to install that wheel we get what we want.
6. Install `repo-info`

Closes: canonical/hotsos#332
Closes: canonical/hotsos#464
Closes: canonical/hotsos#469
Signed-off-by: Nicolas Bock <nicolas.bock@canonical.com>