Improvements to config management
This patch fixes the metaclass implementation of hotsos.config
and make several improvements to the implementation of config
such as making config strictly defined i.e. if code tries to
set or get config that is not recognised it will now raise an
error. Config options are now lowercase in code and all have
a description and default value.