Fix keystone ldap check for LP 1896125
python3-keystone is installed as a dep for all openstack services
so we need to check for package "keystone" to know this is a
keystone host and we also now check that ldap is actually in use.

We also extend the check to ensure that ldap is in-use by asserted
that one or more keystone config file has it configured. This
required extending ycheck requires.types.config to support paths
that are globs as well as lists of paths.