[ceph] Warn if SSD OSDs are using bcache
It's nearly always the case that the SSDs can provide better
performance on their own compared to a bcache device which is
typically shared between several OSDs. Bcache usage is mainly
meant when using OSDs are HHDs. So this is likely to be a
misconfiguration and recommend removing the bcache device(s).

Fixes #573.

Signed-off-by: Ponnuvel Palaniyappan <pponnuvel@gmail.com>