Maximise scenario search grouping
When we run scenariosn we now attempt to pre-load as
many checks searches as possible so as to do them once
in parallel. Also fixes CommonTimestampMatcher to only
fetch year from date() once and only if necessary.

Also fixes some scenario search expressions that were
not correctly formatted to allow the post-processing
constraints to be applied.