Don't cache result groups in checks handler
Result groups are used when formatting messages
but it is more efficient to obtain them on the fly than to
attempt to cache them. This patch also removes support for
search-result-format-groups since it is confusing and
misleading. Intsead we now only have format-dict - docs
updated to reflect this.

Adds a new "first" render function to get the first item
in a list and ensures that result group numbers are not
zero-indexed so that they match actual result groups.