Align timestamp regex expr across searches
This patch aligns the format used to match start of line timestamps
across events and scenarios for all plugins. It also ensures that
(a) every search has at least one subgroup to avoid saving the entire
matched string and (b) ensures that we only use a subgroup for time
if it is actually needed since using it will increase memory usage
due to hh:mm:ss values not being deduplicated.