host_helpers/systemd: fix start_time does not respect timezone when reading from systemd status
the start_time method first tries to pull the start time info from system journal
which already accounts for the timestamp's timezone. if it fails, it falls back to
parsing `systemctl status --all` output and it ignores the specified timezone in the
timestamp, which causes it to treat time in the local timezone of the running system.
this patch fixes that.

Fixes #689

Signed-off-by: Mustafa Kemal Gilor <mustafa.gilor@canonical.com>