Update extra constraints logic
Hotsos uses searchkit to apply constraints on files
it searches e.g. filtering logs older than a certain
date. However there are some filters supported by
the constraints property that are not supported by
SearchConstraintSearchSince so have to be applied
to search results. This removes the filter by age
since that is already supported by
SearchConstraintSearchSince but others remain.