Cleanup temporary files and directories better
Changed the HotSOSClient class to unset
HotSOSConfig.plugin_tmp_dir after running plugins,
else generating a summary could re-create the temp
directory structure based on the name of the plugin
last run.

If there was an exception running any of the plugins,
a message is now printed along with the name of the
temporary log file as it may have useful information.
Otherwise the temp file is removed.

Resolves: #730