Fix spinner message printing multiple times
When analysing a tarfile, attempting to print the
"INFO: analysing $file" line in the spinner code
can overflow the window if it's narrow, resulting
in hundreds of lines of extra output.

Change to print what is being analysed on its own
line to keep the "analysing" spinner message short
and prevent the extra output.

Resolves: #729