Second attempt at temp file and directory cleanup
Ensure HotSOSConfig.plugin_tmp_dir is unset after running
plugins and delete temp log file once we are done with it.

If there was an exception running any of the plugins,
a message is now printed along with the name of the
temporary log file as it may have useful information.
Otherwise the temp file is removed.

Also ensures that if plugin_tmp_dir is None the
host_helpers cache uses a new NullCache so that calls
still work but are not cached. This allows e.g. the CLIhelper
to be used after the plugin env teardown()

Resolves: #730