Fix how we cache check properties
Previously we did not save the cache of logically grouped
items because the semantics around doing that are a little
fuzzy but that meant that scenarios could not reference the
cache of a check that had grouped requirement types. This
patch enables the caching of grouped requirements on the basis
that a cache will be saved if (A) the result is True and
(B) it is the first to be saved in a group or list (C) it
the same has type as the previous one to be saved i.e.
'A AND (B or C)'.