Check for ovn northd version mismatch
If the versions are from the same major release
and ovn-match-northd-version='true' in the local
db then we can suggest setting
ovn-match-northd-version='false' to unblock the
controller. Also adds a scenario to warn if
ovn-match-northd-version is set to true since
this should only be the case during a major release
upgrade.

Resolves: #763