Added two more apparmor profiles
On my system, apparmor_status also reports modes
'kill' and 'unconfined', so add them to the parser.

While testing I noticed the unit test kept matching
the processes section in addition to the profiles
one, leading to more profiles than were in the list.
Changing the regex to be stricter resolved the issue.

TrivialFix