Add huge pages check
The patch adds logic for testing if there are too many huge pages
defined in the system, which are not being fully utilised, but occupy
memory which can't be used for any other purpose. Such situations can
negatively impact performance or even lead to the out of memory
situations.

This patch also introduces a small refactoring of the VMStat class
in the hotsos/core/plugins/kernel/memory.py, in order to avoid code
duplication (between VMStat and MemInfo classes).