Add generic binary requirements checker (#828)
We have requirement types for packages like apt, snap etc
but do not have a way to get information from a binary.
Some applications are not installed by packages but may
still have e.g. a version so having a way to query this
is useful. This tries to provide a common way to do do
this and uses the Juju plugin/binary as an initial
use-case.

Resolves: #815