Catch and verify log calls
There are some tests for invalid data that cause either
log.warning() or log.error() to be called. In order to
not pollute test output (these are shown in RED), mock
the calls and actually verify they were called.

In the case of template-generated tests, just mock the
log call to prevent the output.

TrivialFix