Make common global searcher
First part of changes to make a common global searcher that
can be used for both events and scenarios. This patch also
fixes the double searching of events in openstack
AgentEventChecks.

This patches make use of a context manager that manipulates
a module global that will eventaully be removedin future
patches.