Remove need for module level global searcher (#889)
By allowing the event handlers in
plugin_extensions.openstack.agent.events to run on their own
we remove the final constraint that now allows using a context
manager to manage a global searcher object that is shared
across all parts in a plugin. This is now shared with all
implementations of YHanlderBase i.e. events and scenarios. A
future patch will enable pre-loading scenario searches into
the global searcher.