Fix BFD state change check (#879)
The way we were checking for bfd state transitions was
incorrectly counting the number of actual transitions.
What we really care about is the state transitions
to down. The logic to do this has had to be moved into
code since it is not possible to do it purely in yaml.

This patch implements the first use of the GlobalSearcher
outside of a yaml search property. To achieve this it
moves the extra search constraints logic into
hotsos.core.search so that it is possible to consume
from within and outside the yaml checks code.

Also modifies pylint ignore-long-lines to allow docstrings
with long urls.