hotsos/core: add name aliasing support
at the moment, scenarios are using the long import paths in order to
reference to a Python property. this feature allows assigning an alias
to a Python class or variable in order to make using plugin property
interfaces easier.

Added aliases to the plugins.
Updated existing scenarios to use aliases.
Added unit tests for the aliasing code.
Updated the docs.

Fixes #912

Signed-off-by: Mustafa Kemal Gilor <mustafa.gilor@canonical.com>