Re-enable pylint duplicate-code check (#956)
The bulk of the changes in this patch are to the plugin_extensions
that implement summary output. Most plugins have some entries in
common and those are now moved to a base class so that they don't
need to be redefined by each plugin.

Also optimises some of the hosthelpers and checks implementations
to satisfy the duplicate-code check. Some disables have been used
where it was deemed that fixing the code would introduce too much
instability and this can be addressed in future patches.