Try journald for kernlogs then fallback to files
Searching journald is fastest but not always available
or possible so fall back to kern.log or file-based journal
from sos if not available.

Also pulls kern.log customisation out of the fake_data_root
and puts into the unit test files so that the data root is
restored to its original state.