[ceph] Skip suboptimal PG count warning on low utilisation
When the cluster is empty or utilised at a very low level,
the PG autoscaler won't increase the PGs/OSD to the optimal
level - it'll be done as the cluster hosts more data.
In such cases, the PGs/OSD count will be lower than what we
consider as optimal range (50 - 200). In such cases, this
warning has no benefit and becomes a false positive. So
we only report this if the cluster usage is 10% or more.

Fixes #1016.

Signed-off-by: Ponnuvel Palaniyappan <pponnuvel@gmail.com>